+++ lofty/src/iff/wav/tag/read.rs
@@ -1,7 +1,7 @@
 use super::RiffInfoList;
 use crate::config::ParsingMode;
 use crate::error::{ErrorKind, Result};
-use crate::iff::chunk::Chunks;
+use crate::iff::chunk::{Chunks, valid_fourcc};
 use crate::macros::decode_err;
 use crate::util::text::utf8_decode_str;
 
@@ -10,26 +10,24 @@ use std::io::{Read, Seek};
 use byteorder::LittleEndian;
 
 pub(in crate::iff::wav) fn parse_riff_info<R>(
-	data: &mut R,
-	chunks: &mut Chunks<LittleEndian>,
-	end: u64,
+	chunks: &mut Chunks<&mut R, LittleEndian>,
 	tag: &mut RiffInfoList,
 	parse_mode: ParsingMode,
 ) -> Result<()>
 where
 	R: Read + Seek,
 {
-	while data.stream_position()? != end && matches!(chunks.next(data), Ok(true)) {
-		let key_str = utf8_decode_str(&chunks.fourcc)
+	while let Some(mut chunk) = chunks.next(parse_mode)? {
+		let key_str = utf8_decode_str(&chunk.fourcc)
 			.map_err(|_| decode_err!(Wav, "Invalid item key found in RIFF INFO"))?;
 
-		if !verify_key(key_str) {
+		if !valid_fourcc(chunk.fourcc) {
 			decode_err!(@BAIL Wav, "RIFF INFO item key contains invalid characters");
 		}
 
 		let key = key_str.to_owned();
 		let value;
-		match chunks.read_cstring(data) {
+		match chunk.read_cstring() {
 			Ok(cstr) => value = cstr,
 			Err(e) => {
 				if parse_mode == ParsingMode::Strict {
@@ -51,11 +49,4 @@ where
 	}
 
 	Ok(())
-}
-
-pub(super) fn verify_key(key: &str) -> bool {
-	key.len() == 4
-		&& key
-			.chars()
-			.all(|c| c.is_ascii_uppercase() || c.is_ascii_digit())
 }